Add files via upload
The driving force behind this project was the intention to make a “Problem Solving
Project”. During the examination season, there is a chance of leakage of question
paper. In day-to-day life, many cases are known to be heard about question paper
leakage and due to this, exams are postponed. Hence such incidents have a serious
negative impact on the society. To avoid such leakage of question papers, it is
necessary to provide them security. For this purpose we can implement the system
named as QUESTION PAPER LEAKAGE PREVENTION SMART LOCKER
(LOCATION BASED UNLOCKING). This system helps us to provide the security to
question papers & also used for providing the security to the important documents. If
problems related to exams are avoided using such systems, then we can say that,
education is fairly implemented. ABSTRACT
In the today’s era students often have to face the problem about examination being
postponed/cancelled due to the leakage of question paper. Because the Board or
University will send question papers to exam centre/colleges before the exam starts
and in that process of handling the question papers for the various exams, maximum
cases of question paper leakage take place. To stop this malpractice of leaking the
question paper “QUESTION PAPER LEAKAGE PREVENTION
SMART LOCKER (LOCATION BASED UNLOCKING)” i.e “LOCATION & GSM
BASED SECURITY SYSTEM FOR EXAMINATION PAPER” is developed.GSM
modem is connected to the box containing question papers along with the
microcontroller. The location of the exam centre or college is pre defined in the
system. So when the locker (containing the question papers) reaches the exam centre,
it checks/ compares the current location with the stored location of the exam centre in
the system. If the location matches the locker sends a real-time OTP SMS on an
authorized pre-programmed mobile number (ex: principal’s mobile number). When
this OTP is entered correctly only then the locker opens for a predefined time (in
which you can collect the question papers).If you enter 3 consecutive WRONG OTPs
then the system will get LOCKED! After which nothing can be done unless the
system is RESET. ALSO if Somebody STEALS the locker itself and tries to open it
forcefully (by hitting or hammering it) then it also has a vibration sensor that senses
this hitting/hammering and AGAIN sends a REAL-TIME SMS on the Authorized
mobile number. This SMS will contain the GOOGLE MAPS LINK for the
CURRENT LOCATION of the locker, so that the locker can be traced. WORKING PRINCIPLE:
The location of the exam centre is pre-defined in the system. When the locker
(containing the question papers), reaches the exam centre, it checks for the location (at
every 5 seconds) and if it matches, the system sends an OTP on a pre-defined
authorized mobile number by means of SMS. The OTP is entered using the keyboard
of the system. A correct OTP opens the door of the locker for a predefined interval of
time and then closes it automatically. Continuously 3 wrong OTP entries, locks the
system. ALSO if the locker is stolen or somebody tries to open it forcefully at some
other location by hitting it etc. Then the vibration sensor senses the vibrations and
sends an SMS on the pre-programmed mobile number, This SMS contains the
GOOGLE MAPS LINK for the current location of the locker. OTHER APPLICATIONS & FUTURE IMPROVEMENTS:
1) In ELECTIONS to store EVM and BALLOT machines (so only on reaching the
correct polling booth the locker is opened).
2) Can be used by BANKS for transportation of CASH.
3) Can be used by MILITARY for transportation of CONFIDENTIAL MATERIAL.
4) Personal USE.
5) Question paper leakage prevention.
FUTURE IMPROVEMENTS:
1) SELF powered system
2) LIVE tracking
3) Multiple OTP and OTP sequence
4) Encryption
5) Finger print / IRIS scanner / Facial recognition can be added to the system to
increase the level of security.